+++ CHANGELOG.md
@@ -1,8 +1,14 @@
 Changelog
 =========
 
+## TBD
+
+### Bug fixes
+
 * Support adding pre-delivery metadata to out-of-memory reports
   [#393](https://github.com/bugsnag/bugsnag-cocoa/pull/393)
+* Fix erroneously reporting out-of-memory events from iOS app extensions
+  [#394](https://github.com/bugsnag/bugsnag-cocoa/pull/394)
 
 ## 5.22.3 (2019-07-15)